+++ packages/ui/src/watch/history-page/container/index.test.tsx
@@ -1,11 +1,11 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { render } from '@testing-library/react';
 import { HistoryContainer } from './index';
-import { VideoSkeleton } from '../../videos/video-card-skeleton';
+import { VideoSkeleton } from '../../videos/video-card/skeleton';
 import { VideoCard } from '../../videos/video-card';
 
 // Mock components
-vi.mock('../../videos/video-card-skeleton', () => ({
+vi.mock('../../videos/video-card/skeleton', () => ({
   VideoSkeleton: vi.fn(() => <div>VideoSkeleton</div>),
 }));
